pyproject.toml
[tool.poetry]
name = "sla-exporter"
version = "0.7.0"
description = "Prometheus Exporter for check service healthy and get sla metrics"
authors = ["lfyzjck <jickimkim@gmail.com>"]
license = "GPL-3.0-or-later"


[tool.poetry.dependencies]
python = "^3.6"
prometheus_client = "^0.8.0"
apscheduler = "^3.6.3"
pyyaml = "^5.3.1"
requests = "^2.24.0"
twisted = "^20.3.0"
setuptools = "49.6.0"
schematics = "^2.1.0"
SQLAlchemy = "^1.3.20"
mysqlclient = "^2.0.1"
PyHive = "^0.6.3"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
black = {version = "^20.8b1", allow-prereleases = true}

[tool.poetry.scripts]
sla-exporter = "sla_exporter.main:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
